chore: The great refactor begins
- Added premake as a submodule and the binary is built using a batch script
- Added batch scripts to build premake, the project and other helper stuff
- Implemented vscode premake action to generate tasks.json and launch.json and clean premake task to
clean binary files and generated project files
- The debugging and build experience in VSCode is fairly good:
	- Uses clangd as the LSP
	- premake generates compile_commands.json which clangd uses

+++ core/Xenode/src/core/asset/EditorAssetManager.cpp
@@ -2,6 +2,7 @@
 #include "EditorAssetManager.h"
 
 #include <core/app/Timer.h>
+#include <cstdint>
 #include <project/ProjectManager.h>
 
 #include "AssetImporter.h"
@@ -74,7 +75,7 @@ namespace Xen
 		else
 		{
 			if(handle) // Log only when non zero handles are requested
-				XEN_ENGINE_LOG_ERROR("Handle '{0}' is invalid!", handle);
+				XEN_ENGINE_LOG_ERROR("Handle '{0}' is invalid!", (uint64_t)handle);
 			return nullptr;
 		}
